[BOSK-86] feat: added "predict proba" to RF
