+++ examples/deep_forests/gpu/classification_decision_tree_jax.py
@@ -15,20 +15,20 @@ loaders = {
     "breast_cancer": load_breast_cancer,
 }
 
-from bosk.block.zoo.models.classification.classification_models_jax import DecisionTreeClassifierBlock
+from bosk.block.zoo.models.classification.classification_models_jax import RandomForestClassifierBlock
 
 
-def simple_example():
-    dataset = load_iris()
-    X, y = dataset["data"], dataset["target"]
-    n_classes = np.size(np.bincount(y))
-    model = DecisionTreeClassifierBlock(
-        n_classes=n_classes,
-        max_depth=3,
-        min_samples=1,
-    )
-    fitted_model = model.fit(X, y)
-    print(fitted_model.score(X, y))
+# def simple_example():
+#     dataset = load_iris()
+#     X, y = dataset["data"], dataset["target"]
+#     n_classes = np.size(np.bincount(y))
+#     model = RandomForestClassifier(
+#         n_classes=n_classes,
+#         max_depth=3,
+#         min_samples=1,
+#     )
+#     fitted_model = model.fit(X, y)
+#     print(fitted_model.score(X, y))
 
 import numpy as np
 
@@ -44,19 +44,19 @@ from bosk.pipeline.builder.functional import FunctionalPipelineBuilder
 def make_deep_forest_functional_gpu(executor, **ex_kw):
     b = FunctionalPipelineBuilder()
     X, y = b.Input()(), b.TargetInput()()
-    rf_1 = b.DecisionTreeClassifier(n_classes=2)(X=X, y=y)
+    rf_1 = b.RandomForestClassifier(n_classes=2)(X=X, y=y)
     et_1 = b.ETC(random_state=42)(X=X, y=y)
     et_1 = b.MoveTo("GPU")(X=et_1)
     concat_1 = b.Concat(['X', 'rf_1', 'et_1'])(X=X, rf_1=rf_1, et_1=et_1)
     concat_1 = b.MoveTo("GPU")(X=concat_1)
 
-    rf_2 = b.DecisionTreeClassifier(n_classes=2)(X=concat_1, y=y)
+    rf_2 = b.RandomForestClassifier(n_classes=2)(X=concat_1, y=y)
     et_2 = b.ETC(random_state=42)(X=concat_1, y=y)
     et_2 = b.MoveTo("GPU")(X=et_2)
     concat_2 = b.Concat(['X', 'rf_2', 'et_2'])(X=X, rf_2=rf_2, et_2=et_2)
     concat_2 = b.MoveTo("GPU")(X=concat_2)
 
-    rf_3 = b.DecisionTreeClassifier(n_classes=2)(X=concat_2, y=y)
+    rf_3 = b.RandomForestClassifier(n_classes=2)(X=concat_2, y=y)
     et_3 = b.ETC(random_state=42)(X=concat_2, y=y)
     et_3 = b.MoveTo("GPU")(X=et_3)
     stack_3 = b.Stack(['rf_3', 'et_3'], axis=1)(rf_3=rf_3, et_3=et_3)
